# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Land Record Management System API",
    "description": "REST API for Nepal's Land Record Digitization System with Merkle Tree verification. Supports three user roles: Super Admin (Nepal Sarkar), Malpot Officer, and Citizen.",
    "contact": {
      "name": "Abhishek Ojha",
      "email": "ashishojha258@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/verification/record/{id}": {
      "get": {
        "tags": [
          "Verification"
        ],
        "summary": "Verify a single land record's integrity",
        "operationId": "verifyRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/verification/record/{id}/proof": {
      "get": {
        "tags": [
          "Verification"
        ],
        "summary": "Get Merkle proof path for a record",
        "operationId": "getProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProofStepResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProofStepResponse": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          },
          "position": {
            "type": "string"
          }
        }
      },
      "VerificationResponse": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "integer",
            "format": "int64"
          },
          "kittaNumber": {
            "type": "string"
          },
          "valid": {
            "type": "boolean"
          },
          "computedHash": {
            "type": "string"
          },
          "storedHash": {
            "type": "string"
          },
          "merkleRootHash": {
            "type": "string"
          },
          "merkleProof": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProofStepResponse"
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "description": "Enter your JWT token",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}